requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
argon2-cffi=20.1.0=py39h3811e60_2
async_generator=1.10=py_0
at-spi2-atk=2.38.0=h0630a04_2
at-spi2-core=2.38.0=h0630a04_3
atk-1.0=2.36.0=h3371d22_4
attrs=20.3.0=pyhd3deb0d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
bleach=3.3.0=pyh44b312d_0
boost=1.74.0=py39h5472131_3
boost-cpp=1.74.0=hc6e9bd1_2
bzip2=1.0.8=h7f98852_4
ca-certificates=2020.12.5=ha878542_0
cairo=1.16.0=h6cf1ce9_1008
cairomm=1.12.2=ha770c72_3
cairomm-1.0=1.12.2=h56b4340_3
certifi=2020.12.5=py39hf3d152e_1
cffi=1.14.5=py39he32792d_0
cycler=0.10.0=py_2
dbus=1.13.6=h48d8840_2
decorator=5.0.7=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
epoxy=1.5.5=h7f98852_0
expat=2.3.0=h9c3ff4c_0
fontconfig=2.13.1=hba837de_1005
freetype=2.10.4=h0708190_1
fribidi=1.0.10=h36c2ea0_0
gdk-pixbuf=2.42.6=h04a7f16_0
gettext=0.19.8.1=h0b5b191_1005
glib=2.68.1=h9c3ff4c_0
glib-tools=2.68.1=h9c3ff4c_0
gmp=6.2.1=h58526e2_0
gobject-introspection=1.68.0=py39hcb793ab_1
graph-tool=2.37=py39hc2c1061_1
graphite2=1.3.13=h58526e2_1001
gst-plugins-base=1.18.4=h29181c9_0
gstreamer=1.18.4=h76c114f_0
gtk3=3.24.28=h8879c87_0
harfbuzz=2.8.0=h83ec7ef_1
hicolor-icon-theme=0.17=ha770c72_2
icu=68.1=h58526e2_0
importlib-metadata=4.0.1=py39hf3d152e_0
ipykernel=5.5.3=py39hef51801_0
ipython=7.22.0=py39hef51801_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.3=pyhd3deb0d_0
jedi=0.18.0=py39hf3d152e_2
jinja2=2.11.3=pyh44b312d_0
jpeg=9d=h36c2ea0_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter=1.0.0=py39hf3d152e_6
jupyter_client=6.1.12=pyhd8ed1ab_0
jupyter_console=6.4.0=pyhd8ed1ab_0
jupyter_core=4.7.1=py39hf3d152e_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_widgets=1.0.0=pyhd8ed1ab_1
kiwisolver=1.3.1=py39h1a9c180_1
krb5=1.17.2=h926e7f8_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.35.1=hea4e1c9_2
libblas=3.9.0=8_openblas
libcblas=3.9.0=8_openblas
libclang=11.1.0=default_ha53f305_0
libcups=2.2.12=hf10b501_1
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.10=hcdb4288_3
libffi=3.3=h58526e2_2
libgcc-ng=9.3.0=h2828fa1_19
libgfortran-ng=9.3.0=hff62375_19
libgfortran5=9.3.0=hff62375_19
libglib=2.68.1=h3e27bee_0
libgomp=9.3.0=h2828fa1_19
libiconv=1.16=h516909a_0
liblapack=3.9.0=8_openblas
libllvm11=11.1.0=hf817b99_2
libopenblas=0.3.12=pthreads_h4812303_1
libpng=1.6.37=h21135ba_2
libpq=13.1=hfd2b0eb_2
librsvg=2.50.4=hc3c00ef_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.3.0=h6de172a_19
libtiff=4.2.0=hdc55705_0
libuuid=2.32.1=h7f98852_1000
libwebp-base=1.2.0=h7f98852_2
libxcb=1.13=h7f98852_1003
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.10=h72842e0_4
lz4-c=1.9.3=h9c3ff4c_0
markupsafe=1.1.1=py39h3811e60_3
matplotlib-base=3.4.1=py39h2fa2bec_0
mistune=0.8.4=py39h3811e60_1003
mysql-common=8.0.23=ha770c72_1
mysql-libs=8.0.23=h935591d_1
nbclient=0.5.3=pyhd8ed1ab_0
nbconvert=6.0.7=py39hf3d152e_3
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.2=h58526e2_4
nest-asyncio=1.5.1=pyhd8ed1ab_0
notebook=6.3.0=pyha770c72_1
nspr=4.30=h9c3ff4c_0
nss=3.64=hb5efdd6_0
numpy=1.20.2=py39hdbf815f_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=hf7af979_0
openssl=1.1.1k=h7f98852_0
packaging=20.9=pyh44b312d_0
pandas=1.2.4=py39hde0f152_0
pandas-flavor=0.2.0=py_0
pandoc=2.12=h7f98852_0
pandocfilters=1.4.2=py_1
pango=1.48.4=hb8ff022_0
parso=0.8.2=pyhd8ed1ab_0
pcre=8.44=he1b5a44_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.1.2=py39hf95b381_1
pip=21.0.1=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
prometheus_client=0.10.1=pyhd8ed1ab_0
prompt-toolkit=3.0.18=pyha770c72_0
prompt_toolkit=3.0.18=hd8ed1ab_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pycairo=1.20.0=py39hedcb9fc_1
pycparser=2.20=pyh9f0ad1d_2
pygments=2.8.1=pyhd8ed1ab_0
pygobject=3.40.1=py39he5105b2_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py39hf3d152e_7
pyqt-impl=5.12.3=py39h0fcd23e_7
pyqt5-sip=4.19.18=py39he80948d_7
pyqtchart=5.12=py39h0fcd23e_7
pyqtwebengine=5.12.1=py39h0fcd23e_7
pyrsistent=0.17.3=py39h3811e60_2
python=3.9.2=hffdb5ce_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.9=1_cp39
pytz=2021.1=pyhd8ed1ab_0
pyzmq=22.0.3=py39h37b5a0c_1
qt=5.12.9=hda022c4_4
qtconsole=5.0.3=pyhd8ed1ab_0
qtpy=1.9.0=py_0
readline=8.1=h46c0cb4_0
scipy=1.6.2=py39hee8e79c_0
send2trash=1.5.0=py_0
setuptools=49.6.0=py39hf3d152e_3
sigcpp-2.0=2.10.3=h58526e2_0
six=1.15.0=pyh9f0ad1d_0
sparsehash=2.0.3=hf484d3e_1000
sqlite=3.35.4=h74cdb3f_0
terminado=0.9.4=py39hf3d152e_0
testpath=0.4.4=py_0
tk=8.6.10=h21135ba_1
tornado=6.1=py39h3811e60_1
traitlets=5.0.5=py_0
tzdata=2021a=he74cb21_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.36.2=pyhd3deb0d_0
widgetsnbextension=3.5.1=py39hf3d152e_4
xarray=0.17.0=pyhd8ed1ab_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.0=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
zeromq=4.3.4=h9c3ff4c_0
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h516909a_1010
zstandard=0.15.2=py39h3811e60_0
zstd=1.4.9=ha95c52a_0
